Move all Table widget sources and headers to ObmW
