+++ LICENSES-nonfree
@@ -66,8 +66,8 @@ mosaic-x@ncsa.uiuc.edu.
 ObmW Table widget
 =================
 
-Files: Table.c TableUtil.c Xraw/Table.h Xraw/TableP.h Xraw/TableUtil.h
-       Xraw/ScrolledTable.h Xraw/Frame.h Xraw/FrameP.h
+Files: Table.c Table.h TableP.h TableUtil.c TableUtil.h ScrolledTable.h
+       Separator.c Separator.h
 
 Table widget
 Copyright by Vladimir T. Romanovski
