+++ obm/ObmW/Table.c
@@ -31,11 +31,10 @@ Commercial  usage is  also  possible  with  participation of the author.
 
 #include "Xraw/XrawInit.h"
 #include "Xraw/3d.h"
-#include "Xraw/TableP.h"
-#include "Xraw/TableUtil.h"
-#include "Xraw/Frame.h"
+#include "TableP.h"
+#include "TableUtil.h"
 #include "Xraw/Scrollbar.h"
-#include "Xraw/ScrolledTable.h"
+#include "ScrolledTable.h"
 
 #include <X11/Xaw/AsciiText.h>
 #include <X11/Xaw/Viewport.h>
