+++ obm/ObmW/TableP.h
@@ -21,7 +21,7 @@ Commercial  usage is  also  possible  with  participation of it's author.
  ***********************************************************************/
 
 
-#include "ContainerP.h"
+#include "Xraw/ContainerP.h"
 #include "Table.h"
 #include "TableUtil.h"
